fixes on docs, updated notebooks (#523)
